Moved get user into User orm
